pyproject.toml
[project]
name = "firstproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "fastmcp>=2.13.3",
    "mcp[cli]>=1.22.0",
    "requests>=2.32.5",
]
